multiple small improvements (#184)
