+++ crates/prover/Cargo.toml
@@ -25,7 +25,7 @@ hex.workspace = true
 serde = "1.0"
 serde_json = "1.0"
 serde_stacker = "0.1"
-thiserror = "2.0"
+thiserror.workspace = true
 toml = "0.8"
 
 cudarc = { version = "0.9", optional = true }
