+++ crates/types/Cargo.toml
@@ -9,8 +9,8 @@ version.workspace = true
 
 [dependencies]
 types-base = { path = "base", package = "scroll-zkvm-types-base" }
-types-chunk = { path = "chunk", package = "scroll-zkvm-types-chunk", features = ["scroll-compress-ratio"] }
-types-batch = { path = "batch", package = "scroll-zkvm-types-batch" }
+types-chunk = { path = "chunk", package = "scroll-zkvm-types-chunk", features = ["host"] }
+types-batch = { path = "batch", package = "scroll-zkvm-types-batch", features = ["host"] }
 types-bundle = { path = "bundle", package = "scroll-zkvm-types-bundle" }
 
 openvm-sdk = { workspace = true }
@@ -23,4 +23,3 @@ sbv-primitives.workspace = true
 serde.workspace = true
 base64.workspace = true
 bincode_v1.workspace = true
-c-kzg.workspace = true
