+++ crates/types/batch/Cargo.toml
@@ -9,7 +9,6 @@ version.workspace = true
 
 [dependencies]
 alloy-primitives = { workspace = true }
-bincode = { workspace = true }
 rkyv.workspace = true
 serde.workspace = true
 itertools.workspace = true
@@ -22,11 +21,14 @@ openvm-ecc-guest = { workspace = true, features = ["halo2curves"] }
 openvm-pairing-guest = { workspace = true, features = ["halo2curves"] }
 openvm-pairing = { workspace = true, features = ["bls12_381", "halo2curves"] }
 openvm-sha2.workspace = true
-openvm-sha256-guest.workspace = true
 halo2curves-axiom = "0.7.0"
+
+sbv-primitives = { workspace = true, optional = true }
+c-kzg = { workspace = true, optional = true }
 
 [features]
 default = []
+host = ["dep:sbv-primitives", "dep:c-kzg"]
 
 [dev-dependencies]
 c-kzg = { workspace = true }
