+++ crates/types/batch/src/blob_consistency/mod.rs
@@ -4,9 +4,11 @@ use openvm_pairing::bls12_381::Scalar;
 
 mod constants;
 mod openvm;
-pub use openvm::point_evaluation;
+mod types;
 
-pub use openvm::{ToIntrinsic, kzg_to_versioned_hash, verify_kzg_proof};
+pub use openvm::point_evaluation;
+pub use openvm::{kzg_to_versioned_hash, verify_kzg_proof};
+pub use types::ToIntrinsic;
 
 // Number of bytes in a u256.
 pub const N_BYTES_U256: usize = 32;
