+++ crates/types/batch/src/builder/v7.rs
@@ -29,6 +29,16 @@ fn verify_blob_versioned_hash(
     challenge_digest: B256,
     witness: PointEvalWitness,
 ) {
+    #[cfg(feature = "host")]
+    {
+        use crate::utils::point_eval;
+        let kzg_blob = point_eval::to_blob(blob_bytes);
+        let kzg_commitment = point_eval::blob_to_kzg_commitment(&kzg_blob);
+        assert_eq!(
+            point_eval::get_versioned_hash(&kzg_commitment),
+            blob_versioned_hash
+        );
+    }
     let blob_poly = BlobPolynomial::new(blob_bytes);
     // Barycentric evaluation of blob polynomial.
     let (challenge, evaluation) = blob_poly.evaluate(challenge_digest);
